+++ 놀이터(예제 코드 작성)/jackson/src/main/kotlin/com/banjjoknim/playground/jackson/common/Car.kt
@@ -1,4 +1,4 @@
-package com.banjjoknim.playground.model
+package com.banjjoknim.playground.jackson.common
 
 data class Car(
     val name: String,
