+++ 놀이터(예제 코드 작성)/jackson/src/main/kotlin/com/banjjoknim/playground/jackson/common/Owner.kt
@@ -1,3 +1,3 @@
-package com.banjjoknim.playground.model
+package com.banjjoknim.playground.jackson.common
 
 data class Owner(val name: String, val age: Int)
