+++ 놀이터(예제 코드 작성)/jackson/src/main/kotlin/com/banjjoknim/playground/jackson/common/Secret.kt
@@ -1,4 +1,4 @@
-package com.banjjoknim.playground.model
+package com.banjjoknim.playground.jackson.common
 
 import com.fasterxml.jackson.annotation.JacksonAnnotation
